refactor: move utils functions
