+++ tests/test_evaluation_metric.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from recommender.libs.evaluation import ranking_metrics_at_k
+from libs.utils.evaluation import ranking_metrics_at_k
 
 
 def test_map_ndcg():
